requirements.txt
Django==2.2.8
django-cors-headers==3.0.2
django-filter==2.1.0
django-templated-mail==1.1.1
djangorestframework==3.10.2
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==4.3.0
djoser==2.0.1
mercadopago==0.3.5
Pillow==6.2.0
psycopg2==2.8.3
PyJWT==1.7.1
pytz==2019.1
qrcode==6.1
rest-social-auth==2.2.0
sqlparse==0.3.0
django-channels==0.7.0
channels==2.3.1
channels-redis==2.4.1
